+++ ChangeLog.md
@@ -12,6 +12,7 @@
     - Validate the user account structure before creation on server side
     - Remove one success message on user creation, to keep stuck with only one
     - Introduce new AccountsUI.preferredLabel() method and corresponding preferredLabel configuration option
+    - Upgrade pwix:modal version requirement to get Modal export
 
 ### 1.2.5